add upbit, upbit_id and upbit_sg
